+++ packages/cli/src/lib/implementation/core-config.middleware.unit.test.ts
@@ -4,6 +4,7 @@ import { coreConfigMiddleware } from './core-config.middleware';
 import { CoreConfigCliOptions } from './core-config.model';
 import { GeneralCliOptions } from './global.model';
 import { OnlyPluginsOptions } from './only-plugins.model';
+import { SkipPluginsOptions } from './skip-plugins.model';
 
 vi.mock('@code-pushup/core', async () => {
   const { CORE_CONFIG_MOCK }: typeof import('@code-pushup/test-utils') =
@@ -19,7 +20,10 @@ vi.mock('@code-pushup/core', async () => {
 describe('coreConfigMiddleware', () => {
   it('should attempt to load code-pushup.config.(ts|mjs|js) by default', async () => {
     await coreConfigMiddleware(
-      {} as GeneralCliOptions & CoreConfigCliOptions & OnlyPluginsOptions,
+      {} as GeneralCliOptions &
+        CoreConfigCliOptions &
+        OnlyPluginsOptions &
+        SkipPluginsOptions,
     );
     expect(autoloadRc).toHaveBeenCalled();
   });
@@ -27,7 +31,7 @@ describe('coreConfigMiddleware', () => {
   it('should directly attempt to load passed config', async () => {
     await coreConfigMiddleware({
       config: 'cli/custom-config.mjs',
-    } as GeneralCliOptions & CoreConfigCliOptions & OnlyPluginsOptions);
+    } as GeneralCliOptions & CoreConfigCliOptions & OnlyPluginsOptions & SkipPluginsOptions);
     expect(autoloadRc).not.toHaveBeenCalled();
     expect(readRcByPath).toHaveBeenCalledWith(
       'cli/custom-config.mjs',
@@ -38,7 +42,7 @@ describe('coreConfigMiddleware', () => {
   it('should forward --tsconfig option to config autoload', async () => {
     await coreConfigMiddleware({
       tsconfig: 'tsconfig.base.json',
-    } as GeneralCliOptions & CoreConfigCliOptions & OnlyPluginsOptions);
+    } as GeneralCliOptions & CoreConfigCliOptions & OnlyPluginsOptions & SkipPluginsOptions);
     expect(autoloadRc).toHaveBeenCalledWith('tsconfig.base.json');
   });
 
@@ -46,7 +50,7 @@ describe('coreConfigMiddleware', () => {
     await coreConfigMiddleware({
       config: 'apps/website/code-pushup.config.ts',
       tsconfig: 'apps/website/tsconfig.json',
-    } as GeneralCliOptions & CoreConfigCliOptions & OnlyPluginsOptions);
+    } as GeneralCliOptions & CoreConfigCliOptions & OnlyPluginsOptions & SkipPluginsOptions);
     expect(readRcByPath).toHaveBeenCalledWith(
       'apps/website/code-pushup.config.ts',
       'apps/website/tsconfig.json',