+++ packages/nx-plugin/src/executors/internal/context.ts
@@ -1,0 +1,21 @@
+import { ExecutorContext } from 'nx/src/config/misc-interfaces';
+import { BaseNormalizedExecutorContext } from './types';
+
+export type NormalizedExecutorContext = BaseNormalizedExecutorContext & {
+  projectName: string;
+};
+
+export function normalizeContext(
+  context: ExecutorContext,
+): NormalizedExecutorContext {
+  const {
+    projectName = '',
+    root: workspaceRoot,
+    projectsConfigurations,
+  } = context;
+  return {
+    projectName,
+    projectConfig: projectsConfigurations?.projects[projectName],
+    workspaceRoot,
+  };
+}