+++ packages/nx-plugin/src/generators/configuration/schema.d.ts
@@ -1,3 +1,6 @@
-export type ConfigurationGeneratorSchema = {
+export type ConfigurationGeneratorOptions = {
   project: string;
+  targetName?: string;
+  skipTarget?: boolean;
+  skipConfig?: boolean;
 };