+++ packages/plugin-coverage/src/lib/nx/coverage-paths.unit.test.ts
@@ -1,3 +1,5 @@
+import type { JestExecutorOptions } from '@nx/jest/src/executors/jest/schema';
+import type { VitestExecutorOptions } from '@nx/vite/executors';
 import { vol } from 'memfs';
 import { join } from 'node:path';
 import { describe, expect, it } from 'vitest';
@@ -49,6 +51,11 @@ vi.mock('bundle-require', () => ({
       coverageDirectory: 'coverage',
     };
 
+    const JEST_PRESET: JestCoverageConfig & { preset?: string } = {
+      preset: '../../jest.preset.ts',
+      coverageDirectory: 'coverage',
+    };
+
     const wrapReturnValue = (
       value: VitestCoverageConfig | JestCoverageConfig,
     ) => ({ mod: { default: value } });
@@ -67,6 +74,8 @@ vi.mock('bundle-require', () => ({
         return wrapReturnValue(JEST_NO_LCOV);
       case 'jest-no-dir':
         return wrapReturnValue(JEST_NO_DIR);
+      case 'jest-preset':
+        return wrapReturnValue(JEST_PRESET);
       default:
         return wrapReturnValue({});
     }
@@ -78,8 +87,8 @@ describe('getCoveragePathForTarget', () => {
     vol.fromJSON(
       {
         // values come from bundle-require mock above
-        'vitest-valid.config.unit.ts': '',
-        'jest-valid.config.unit.ts': '',
+        'vitest-valid.config.ts': '',
+        'jest-valid.config.ts': '',
       },
       MEMFS_VOLUME,
     );
@@ -89,11 +98,18 @@ describe('getCoveragePathForTarget', () => {
     await expect(
       getCoveragePathsForTarget(
         {
-          executor: '@nx/vite:test',
-          options: { config: 'vitest-valid.config.unit.ts' },
+          name: 'cli',
+          root: join('packages', 'cli'),
+          targets: {
+            test: {
+              executor: '@nx/vite:test',
+              options: {
+                configFile: 'vitest-valid.config.ts',
+              } satisfies VitestExecutorOptions,
+            },
+          },
         },
-        join('packages', 'cli'),
-        'unit-test in cli',
+        'test',
       ),
     ).resolves.toEqual({
       pathToProject: join('packages', 'cli'),
@@ -105,11 +121,18 @@ describe('getCoveragePathForTarget', () => {
     await expect(
       getCoveragePathsForTarget(
         {
-          executor: '@nx/jest:jest',
-          options: { config: 'jest-valid.config.unit.ts' },
+          name: 'core',
+          root: join('packages', 'core'),
+          targets: {
+            test: {
+              executor: '@nx/jest:jest',
+              options: {
+                jestConfig: 'jest-valid.config.ts',
+              } satisfies JestExecutorOptions,
+            },
+          },
         },
-        join('packages', 'core'),
-        'unit-test in core',
+        'test',
       ),
     ).resolves.toBe(join('packages', 'core', 'coverage', 'core', 'lcov.info'));
   });
@@ -118,11 +141,16 @@ describe('getCoveragePathForTarget', () => {
     await expect(() =>
       getCoveragePathsForTarget(
         {
-          executor: '@nx/cypress',
-          options: { config: 'cypress.config.ts' },
+          name: 'ui',
+          root: join('apps', 'ui'),
+          targets: {
+            'component-test': {
+              executor: '@nx/cypress',
+              options: { cypressConfig: 'cypress.config.ts' },
+            },
+          },
         },
-        join('apps', 'ui'),
-        'component-test in ui',
+        'component-test',
       ),
     ).rejects.toThrow('Unsupported executor @nx/cypress.');
   });
@@ -141,12 +169,12 @@ describe('getCoveragePathForVitest', () => {
     );
   });
 
-  it('should fetch Vitest reportsDirectory', async () => {
+  it('should fetch Vitest reportsDirectory from vite config', async () => {
     await expect(
       getCoveragePathForVitest(
-        'vitest-valid.config.unit.ts',
-        join('packages', 'cli'),
-        'unit-test in cli',
+        { configFile: 'vitest-valid.config.unit.ts' },
+        { name: 'cli', root: join('packages', 'cli') },
+        'unit-test',
       ),
     ).resolves.toEqual({
       pathToProject: join('packages', 'cli'),
@@ -157,23 +185,61 @@ describe('getCoveragePathForVitest', () => {
   it('should throw when reportsDirectory is not set in Vitest config', async () => {
     await expect(() =>
       getCoveragePathForVitest(
-        'vitest-no-dir.config.integration.ts',
-        'packages',
-        'integration-test in cli',
+        { configFile: 'vitest-no-dir.config.integration.ts' },
+        { name: 'cli', root: join('packages', 'cli') },
+        'integration-test',
       ),
     ).rejects.toThrow(
       /configuration .* does not include coverage path .* Add the path under coverage > reportsDirectory/,
     );
   });
 
+  it('should override vitest config reportsDirectory from project.json', async () => {
+    await expect(
+      getCoveragePathForVitest(
+        {
+          configFile: 'vitest-valid.config.unit.ts',
+          reportsDirectory: join(
+            '..',
+            '..',
+            'dist',
+            'packages',
+            'cli',
+            'coverage',
+          ),
+        },
+        { name: 'cli', root: join('packages', 'cli') },
+        'unit-test',
+      ),
+    ).resolves.toEqual({
+      pathToProject: join('packages', 'cli'),
+      resultsPath: join('dist', 'packages', 'cli', 'coverage', 'lcov.info'),
+    } satisfies CoverageResult);
+  });
+
   it('should throw when Vitest config does not include lcov reporter', async () => {
     await expect(() =>
       getCoveragePathForVitest(
-        'vitest-no-lcov.config.integration.ts',
-        'packages',
-        'integration-test in core',
+        { configFile: 'vitest-no-lcov.config.integration.ts' },
+        { name: 'core', root: join('packages', 'core') },
+        'integration-test',
       ),
     ).rejects.toThrow(/configuration .* does not include LCOV report format/);
+  });
+
+  it('should handle absolute path in reportsDirectory', async () => {
+    await expect(
+      getCoveragePathForVitest(
+        {
+          configFile: 'vitest-valid.config.unit.ts',
+          reportsDirectory: join(process.cwd(), 'coverage', 'packages', 'cli'),
+        },
+        { name: 'cli', root: join('packages', 'cli') },
+        'unit-test',
+      ),
+    ).resolves.toBe(
+      join(process.cwd(), 'coverage', 'packages', 'cli', 'lcov.info'),
+    );
   });
 });
 
@@ -193,32 +259,116 @@ describe('getCoveragePathForJest', () => {
   it('should fetch Jest coverageDirectory', async () => {
     await expect(
       getCoveragePathForJest(
-        'jest-valid.config.unit.ts',
-        join('packages', 'cli'),
-        'unit-test in core',
+        { jestConfig: 'jest-valid.config.unit.ts' },
+        { name: 'cli', root: join('packages', 'cli') },
+        'unit-test',
       ),
     ).resolves.toBe(join('packages', 'cli', 'coverage', 'core', 'lcov.info'));
   });
 
-  it('should throw when reportsDirectory is not set in Jest config', async () => {
+  it('should throw when coverageDirectory is not set in Jest config', async () => {
     await expect(() =>
       getCoveragePathForJest(
-        'jest-no-dir.config.integration.ts',
-        'packages',
-        'integration-test in core',
+        { jestConfig: 'jest-no-dir.config.integration.ts' },
+        { name: 'core', root: join('packages', 'core') },
+        'integration-test',
       ),
     ).rejects.toThrow(
       /configuration .* does not include coverage path .* Add the path under coverageDirectory/,
     );
   });
 
+  it('should override jest config coverageDirectory from project.json', async () => {
+    await expect(
+      getCoveragePathForJest(
+        {
+          jestConfig: 'jest-valid.config.unit.ts',
+          coverageDirectory: join(
+            '..',
+            '..',
+            'dist',
+            'packages',
+            'cli',
+            'coverage',
+          ),
+        },
+        { name: 'cli', root: join('packages', 'cli') },
+        'unit-test',
+      ),
+    ).resolves.toBe(join('dist', 'packages', 'cli', 'coverage', 'lcov.info'));
+  });
+
   it('should throw when Jest config does not include lcov reporter', async () => {
     await expect(() =>
       getCoveragePathForJest(
-        'jest-no-lcov.config.integration.ts',
-        'packages',
-        'integration-test in core',
+        { jestConfig: 'jest-no-lcov.config.integration.ts' },
+        { name: 'core', root: join('packages', 'core') },
+        'integration-test',
       ),
     ).rejects.toThrow(/configuration .* does not include LCOV report format/);
+  });
+
+  it('should not throw if lcov reporter in project.json instead of jest config', async () => {
+    await expect(
+      getCoveragePathForJest(
+        {
+          jestConfig: 'jest-no-lcov.config.integration.ts',
+          coverageReporters: ['lcov'],
+        },
+        { name: 'core', root: join('packages', 'core') },
+        'integration-test',
+      ),
+    ).resolves.toBeTypeOf('string');
+  });
+
+  it('should throw if lcov reporter from jest config overridden in project.json', async () => {
+    await expect(
+      getCoveragePathForJest(
+        {
+          jestConfig: 'jest-valid.config.integration.ts',
+          coverageReporters: ['text', 'html'],
+        },
+        { name: 'core', root: join('packages', 'core') },
+        'integration-test',
+      ),
+    ).rejects.toThrow(/configuration .* does not include LCOV report format/);
+  });
+
+  it('should not throw if lcov reporter in both project.json and jest config', async () => {
+    await expect(
+      getCoveragePathForJest(
+        {
+          jestConfig: 'jest-valid.config.integration.ts',
+          coverageReporters: ['lcov'],
+        },
+        { name: 'core', root: join('packages', 'core') },
+        'integration-test',
+      ),
+    ).resolves.toBeTypeOf('string');
+  });
+
+  it('should not throw regarding missing lcov reporter if jest config uses preset', async () => {
+    await expect(
+      getCoveragePathForJest(
+        { jestConfig: 'jest-preset.config.ts' },
+        { name: 'core', root: join('packages', 'core') },
+        'test',
+      ),
+    ).resolves.toBeTypeOf('string');
+  });
+
+  it('should handle absolute path in coverageDirectory', async () => {
+    await expect(
+      getCoveragePathForJest(
+        {
+          jestConfig: 'jest-valid.config.unit.ts',
+          coverageDirectory: join(process.cwd(), 'coverage', 'packages', 'cli'),
+        },
+        { name: 'cli', root: join('packages', 'cli') },
+        'unit-test',
+      ),
+    ).resolves.toBe(
+      join(process.cwd(), 'coverage', 'packages', 'cli', 'lcov.info'),
+    );
   });
 });