+++ packages/utils/mocks/fixtures/valid-cjs-module-export.cjs
@@ -1,0 +1,1 @@
+module.exports = 'valid-cjs-module-export';